separate processConnectAck() and regular processMsg() calls
Now the code is more readable and the bad case that we change from
state connected to disconnected within the loop doesn't crash us.

This Patch implies that we don't check the first regular messages
automatically on eConnect anymore. This doesn't breaks anything and
I think it's an advantage that client programmer will not get any
callbacks (except error()) while he is calling eConnect().