fix, connect fails with non blocking sockets on win32
On win32 connect does not set errno and moreover it does return
(WSA)EWOULDBLOCK instead of (WSA)EINPROGRESS. We rewrite it to let it behave
as we want for the cases we need.

Note we haven't noticed this bug in practice because of another bug we
had never used non blocking sockets on win32 so far.