minor, allow to initiate handshake and goodbye procedures on user-provided sockets
This changeset provides the methods handshake() and wavegoodbye() in the posix client socket.
The responsibility to connect to the TWS now lies with the user in exchange for greater
control over the socket.

handshake() must be called 3 times.  Once to send off the handshake token (ideally when
the socket is writable), once to receive the servers acknowledgement (socket should be
readable of course) and once to set the client id.

The wavegoodbye() is comparatively simple.  It will disassociate internal resources from
the socket and forget about the socket altogether.

Signed-off-by: Sebastian Freundt <freundt@ga-group.nl>