minor, change procedure to go through the handshake:
Now handshake() upon error returns -1.  handshake() when finished returns 1 (so handshakeComplete() has gone)
and handshake() in the process of shaking hands returns 0.
That way the handshake process can be implemented as:

while ((rc = handshake()) == 0);

and if rc comes back as 1 the handshake was successful, or as -1, then there
was an error on the way.

Signed-off-by: Sebastian Freundt <freundt@ga-group.nl>